requirements.txt
streamlit
openai
python-dotenv
PyMuPDF
python-docx
